fix ci

+++ padiff/utils.py
@@ -232,9 +232,8 @@ def reset_log_dir():
 
 
 def clean_log_dir():
-    os.rmdir(diff_log_path)
-    # if not os.listdir(diff_log_path):
-    #     os.rmdir(diff_log_path)
+    if not os.listdir(diff_log_path):
+        os.rmdir(diff_log_path)
 
 
 def tensors_mean(inp, mode):
